Refactor, implement in-memory definitions cache
- A lot of the ClientUtils stuff have been moved to their own classes. Checksum stuff is now in ChecksumUtils, Launcher stuff is now in LauncherUtils, etc...

- Definitions are now cached in-memory for up-to a day or when the game stops or crashes - whichever is first. This improves performance and reduces bandwidth usage, especially on busy servers.
