+++ src/main/java/dk/zlepper/itlt/common/ChecksumType.java
@@ -1,0 +1,7 @@
+package dk.zlepper.itlt.common;
+
+public enum ChecksumType {
+    BLAKE3_256,
+    BLAKE3_224,
+    Default
+}
